Move EqualityComparerTests
